feat: introduce new incoming message parser pattern (#1340)
* feat: introduce new incoming message parser pattern
