+++ source/BuildingBlocks.Application/FeatureFlag/FeatureFlagName.cs
@@ -39,4 +39,9 @@ public enum FeatureFlagName
     /// Whether to use orchestration for handling RequestWholesaleServices processes.
     /// </summary>
     UseRequestWholesaleServicesProcessOrchestration,
+
+    /// <summary>
+    /// Whether to use the new incoming message parser.
+    /// </summary>
+    UseNewIncomingMessageParser,
 }
