+++ source/BuildingBlocks.Application/FeatureFlag/MicrosoftFeatureFlagManager.cs
@@ -37,5 +37,7 @@ public class MicrosoftFeatureFlagManager : IFeatureFlagManager
     public Task<bool> UseRequestWholesaleServicesProcessOrchestrationAsync() =>
         IsEnabledAsync(FeatureFlagName.UseRequestWholesaleServicesProcessOrchestration);
 
+    public Task<bool> UseNewIncomingMessageParserAsync() => IsEnabledAsync(FeatureFlagName.UseNewIncomingMessageParser);
+
     private Task<bool> IsEnabledAsync(FeatureFlagName featureFlagName) => _featureManager.IsEnabledAsync(featureFlagName.ToString());
 }
