+++ source/BuildingBlocks.Tests/TestDoubles/FeatureFlagManagerStub.cs
@@ -30,4 +30,6 @@ public class FeatureFlagManagerStub : IFeatureFlagManager
     public Task<bool> ReceiveMeteredDataForMeasurementPointsAsync() => Task.FromResult(true);
 
     public Task<bool> UseRequestWholesaleServicesProcessOrchestrationAsync() => Task.FromResult(false);
+
+    public Task<bool> UseNewIncomingMessageParserAsync() => Task.FromResult(true);
 }
