+++ source/IncomingMessages.Application/UseCases/ReceiveIncomingMarketMessage.cs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using BuildingBlocks.Application.FeatureFlag;
 using Energinet.DataHub.EDI.ArchivedMessages.Interfaces;
 using Energinet.DataHub.EDI.ArchivedMessages.Interfaces.Models;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Authentication;
@@ -31,6 +32,8 @@ namespace Energinet.DataHub.EDI.IncomingMessages.Application.UseCases;
 public class ReceiveIncomingMarketMessage
 {
     private readonly MarketMessageParser _marketMessageParser;
+    private readonly IDictionary<(IncomingDocumentType, DocumentFormat), IMessageParser> _messageParsers;
+    private readonly IFeatureFlagManager _featureFlagManager;
     private readonly ValidateIncomingMessage _validateIncomingMessage;
     private readonly ResponseFactory _responseFactory;
     private readonly IArchivedMessagesClient _archivedMessagesClient;
@@ -42,6 +45,8 @@ public class ReceiveIncomingMarketMessage
 
     public ReceiveIncomingMarketMessage(
         MarketMessageParser marketMessageParser,
+        IDictionary<(IncomingDocumentType DocumentType, DocumentFormat DocumentFormat), IMessageParser> messageParsers,
+        IFeatureFlagManager featureFlagManager,
         ValidateIncomingMessage validateIncomingMessage,
         ResponseFactory responseFactory,
         IArchivedMessagesClient archivedMessagesClient,
@@ -52,6 +57,8 @@ public class ReceiveIncomingMarketMessage
         AuthenticatedActor actorAuthenticator)
     {
         _marketMessageParser = marketMessageParser;
+        _messageParsers = messageParsers;
+        _featureFlagManager = featureFlagManager;
         _validateIncomingMessage = validateIncomingMessage;
         _responseFactory = responseFactory;
         _archivedMessagesClient = archivedMessagesClient;
@@ -71,11 +78,12 @@ public class ReceiveIncomingMarketMessage
     {
         ArgumentNullException.ThrowIfNull(documentType);
         ArgumentNullException.ThrowIfNull(incomingMarketMessageStream);
-
-        var incomingMarketMessageParserResult =
-            await _marketMessageParser.ParseAsync(incomingMarketMessageStream, incomingDocumentFormat, documentType, cancellationToken)
-                .ConfigureAwait(false);
-
+        var incomingMarketMessageParserResult = await ParseIncomingMessageAsync(
+                incomingMarketMessageStream,
+                incomingDocumentFormat,
+                documentType,
+                cancellationToken)
+            .ConfigureAwait(false);
         if (incomingMarketMessageParserResult.Errors.Count != 0
             || incomingMarketMessageParserResult.IncomingMessage == null)
         {
@@ -137,6 +145,24 @@ public class ReceiveIncomingMarketMessage
     private static bool ShouldArchive(IncomingDocumentType documentType)
     {
         return documentType != IncomingDocumentType.MeteredDataForMeasurementPoint;
+    }
+
+    private async Task<IncomingMarketMessageParserResult> ParseIncomingMessageAsync(
+        IIncomingMarketMessageStream incomingMarketMessageStream,
+        DocumentFormat documentFormat,
+        IncomingDocumentType documentType,
+        CancellationToken cancellationToken)
+    {
+        if (await _featureFlagManager.UseNewIncomingMessageParserAsync().ConfigureAwait(false))
+        {
+            if (_messageParsers.TryGetValue((documentType, documentFormat), out var messageParser))
+            {
+                return await messageParser.ParseAsync(incomingMarketMessageStream, cancellationToken).ConfigureAwait(false);
+            }
+        }
+
+        return await _marketMessageParser.ParseAsync(incomingMarketMessageStream, documentFormat, documentType, cancellationToken)
+            .ConfigureAwait(false);
     }
 
     private async Task ArchiveIncomingMessageAsync(
